backend/pyproject.toml
[tool.poetry]
name = "socratic-tutor-backend"
version = "0.1.0"
description = "AI Socratic Tutor Backend API"
authors = ["Mohammed <mohammed@equitechfutures.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.0"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.31.0"
pydantic = {extras = ["email"], version = "^2.12.5"}
pydantic-settings = "^2.6.0"
python-dotenv = "^1.0.0"
anthropic = "^0.40.0"
openai = "^1.54.0"
pgvector = "^0.3.5"
python-multipart = "^0.0.17"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
supabase = "^2.10.0"
greenlet = "^3.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
black = "^24.10.0"
ruff = "^0.8.0"
httpx = "^0.27.0"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = "-v --cov=app --cov-report=term-missing"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
